Listen for onBlockedClientConnecting whenever callback is registered
