+++ mobile/src/main/java/be/mygod/vpnhotspot/manage/TetherManager.kt
@@ -2,7 +2,6 @@ package be.mygod.vpnhotspot.manage
 
 import android.Manifest
 import android.annotation.TargetApi
-import android.content.ClipData
 import android.content.Context
 import android.content.Intent
 import android.content.pm.PackageManager
@@ -174,21 +173,6 @@ sealed class TetherManager(protected val parent: TetheringFragment) : Manager(),
         override fun onCapabilityChanged(capability: Parcelable) {
             this.capability = capability
             data.notifyChange()
-        }
-        @RequiresApi(30)
-        override fun onBlockedClientConnecting(client: Parcelable, blockedReason: Int) {
-            @Suppress("NAME_SHADOWING")
-            val client = WifiClient(client)
-            val macAddress = client.macAddress
-            var name = macAddress.toString()
-            if (BuildCompat.isAtLeastS()) client.apInstanceIdentifier?.let { name += "%$it" }
-            val reason = WifiApManager.clientBlockLookup(blockedReason, true)
-            Timber.i("$name blocked from connecting: $reason ($blockedReason)")
-            SmartSnackbar.make(parent.getString(R.string.tethering_manage_wifi_client_blocked, name, reason)).apply {
-                action(R.string.tethering_manage_wifi_copy_mac) {
-                    app.clipboard.setPrimaryClip(ClipData.newPlainText(null, macAddress.toString()))
-                }
-            }.show()
         }
 
         override val title get() = parent.getString(R.string.tethering_manage_wifi)
